server code ddd
