+++ src/product/type.ts
@@ -1,0 +1,3 @@
+import type { RouterOutputs } from "@/trpc";
+
+export type ReviewData = NonNullable<RouterOutputs["user"]["review"]["getOne"]>;
