+++ src/product/validation.ts
@@ -1,0 +1,1 @@
+export const barcodeMinLength = 5;
