+++ src/server/api/router.ts
@@ -1,0 +1,10 @@
+import { productRouter } from "./product";
+import { createTRPCRouter } from "./trpc";
+import { userRouter } from "./user";
+
+export const apiRouter = createTRPCRouter({
+  product: productRouter,
+  user: userRouter,
+});
+
+export type ApiRouter = typeof apiRouter;
