+++ src/server/api/utils/pagination.ts
@@ -10,16 +10,20 @@ const cursorBaseSchema = z
   .refine(isSome, "Supplied string is an invalid base64 encoded json")
   .transform((option) => option.value);
 
-export function createPagination<Z extends Zod.Schema, S extends string>(
-  cursor: Z,
-  sortCols: NonEmptyArray<S>,
-) {
+type PaginationOptions<Z extends Zod.Schema, S extends string> = {
+  cursor: Z;
+  sortBy: NonEmptyArray<S>;
+};
+export function createPagination<Z extends Zod.Schema, S extends string>({
+  cursor,
+  sortBy,
+}: PaginationOptions<Z, S>) {
   return {
     schema: z.object({
       cursor: cursorBaseSchema.pipe(cursor).optional(),
       limit: limitBaseSchema,
       sort: z.object({
-        by: z.enum(sortCols),
+        by: z.enum(sortBy),
         desc: z.boolean(),
       }),
     }),
