+++ src/server/redis/index.ts
@@ -1,0 +1,7 @@
+import { env } from "@/env/index.mjs";
+import { Redis } from "@upstash/redis";
+
+export const redis = new Redis({
+  url: env.UPSTASH_REDIS_REST_URL,
+  token: env.UPSTASH_REDIS_REST_TOKEN,
+});
