+++ src/server/uploadthing/api.ts
@@ -1,0 +1,6 @@
+import { env } from "@/env/index.mjs";
+import { createUploadthing } from "uploadthing/next-legacy";
+import { UTApi } from "uploadthing/server";
+
+export const uploadthing = createUploadthing();
+export const utapi = new UTApi({ apiKey: env.UPLOADTHING_SECRET });
